Add createConnection API to Client.
createConnection is a parameter that lets the user supply their own method of creating a socket to
the LDAP server, as an alternative for the existing logic, which uses net.connect or tls.connect
depending on whether the URL is ldaps or ldap.

The primary use case I have in mind is connecting to an LDAP server through an explicit proxy which
uses the HTTP CONNECT method. In this case, the user can supply a createConnection function handles
making the HTTP CONNECT request and (for ldaps) doing a TLS handshake over the resulting connection.