Listens on all address when host is omitted
Actually when using listen and host is omitted the server will accept connections directed to any IPv4 address but not to IPv6 addresses.

With this patch the server will accept connections on the unspecified IPv6 address (::) when IPv6 is available, or the unspecified IPv4 address (0.0.0.0) otherwise.

Note: In most operating systems, listening to the unspecified IPv6 address (::) may cause the server to also listen on the unspecified IPv4 address (0.0.0.0).